+++ src/komponenter/footer/footer-regular/footer-topp/FooterTopp.tsx
@@ -12,6 +12,7 @@ import Arbeidsflatevalg from './arbeidsflatevalg/Arbeidsflatevalg';
 import { LinksLoader } from '../../../common/content-loaders/LinkLoader';
 import FooterLenker from 'komponenter/footer/common/Lenker';
 import { Locale } from 'store/reducers/language-duck';
+import { DelSkjermLenke } from 'komponenter/footer/common/del-skjerm-lenke/DelSkjermLenke';
 import './FooterTopp.less';
 
 const FooterTopp = () => {
@@ -19,8 +20,10 @@ const FooterTopp = () => {
     const { language } = useSelector((state: AppState) => state.language);
     const context = useSelector((state: AppState) => state.arbeidsflate.status);
     const { data } = useSelector((state: AppState) => state.menypunkt);
-
     const [columnsNode, settColumnsNode] = useState<MenyNode>();
+    const [personvernNode, settPersonvernNode] = useState<MenyNode>();
+    const { PARAMS } = useSelector((state: AppState) => state.environment);
+
     useEffect(() => {
         const languageNode = getLanguageNode(language, data);
         const isLanguageNorwegian = language === Locale.BOKMAL || language === Locale.NYNORSK;
@@ -40,6 +43,13 @@ const FooterTopp = () => {
         }
     }, [language, context, data, settColumnsNode]);
 
+    useEffect(() => {
+        const noder = getLanguageNode(language, data);
+        if (noder && !personvernNode) {
+            settPersonvernNode(findNode(noder, 'Personvern'));
+        }
+    }, [data, personvernNode]);
+
     const scrollToTop = (event: React.MouseEvent) => {
         event.preventDefault();
         window.scrollTo({
@@ -52,38 +62,63 @@ const FooterTopp = () => {
 
     return (
         <div className={cls.className}>
+            <div className="menylenker-seksjon til-toppen">
+                <div className="til-toppen-innhold">
+                    <LenkeMedIkon
+                        onClick={scrollToTop}
+                        tekst={<Tekst id="footer-til-toppen" />}
+                        ikon={<Up />}
+                        venstrestiltIkon={true}
+                        id="footer-til-toppen"
+                    />
+                </div>
+            </div>
             <div className="topp-kolonner">
-                <div className="menylenker-seksjon til-toppen">
-                    <div className="til-toppen-innhold">
-                        <LenkeMedIkon
-                            onClick={scrollToTop}
-                            tekst={<Tekst id="footer-til-toppen" />}
-                            ikon={<Up />}
-                            venstrestiltIkon={true}
-                            id="footer-til-toppen"
-                        />
+                <div className="venstre">
+                    {columnsNode
+                        ? columnsNode.children.slice(0, 2).map((columnNode, i) => (
+                              <div key={i} className={'menylenker-seksjon'}>
+                                  <Heading level="2" size="small" className="menylenker-overskrift">
+                                      {columnNode.displayName}
+                                  </Heading>
+                                  <ul>
+                                      <FooterLenker node={columnNode} />
+                                  </ul>
+                              </div>
+                          ))
+                        : [...Array(3)].map((_, index) => (
+                              <div className={'menylenker-seksjon'} key={index}>
+                                  <LinksLoader id={`footer-link-loader-${index}`} />
+                              </div>
+                          ))}
+                </div>
+                <div className="høyre">
+                    {columnsNode
+                        ? columnsNode.children.slice(2, 4).map((columnNode, i) => (
+                              <div key={i} className={'menylenker-seksjon'}>
+                                  <Heading level="2" size="small" className="menylenker-overskrift">
+                                      {columnNode.displayName}
+                                  </Heading>
+                                  <ul>
+                                      <FooterLenker node={columnNode} />
+                                  </ul>
+                              </div>
+                          ))
+                        : [...Array(3)].map((_, index) => (
+                              <div className={'menylenker-seksjon'} key={index}>
+                                  <LinksLoader id={`footer-link-loader-${index}`} />
+                              </div>
+                          ))}
+                </div>
+                <div className="høyre">
+                    <Arbeidsflatevalg />
+                    <div className={cls.element('bottom-lenker')}>
+                        <ul className={cls.element('personvern-lenker')}>
+                            <FooterLenker node={personvernNode} />
+                            {PARAMS.SHARE_SCREEN && <DelSkjermLenke />}
+                        </ul>
                     </div>
                 </div>
-                {columnsNode
-                    ? columnsNode.children.map((columnNode, i) => (
-                          <div key={i} className={`menylenker-seksjon ${!i ? 'venstre' : i === 2 ? 'hoyre' : 'midt'}`}>
-                              <Heading level="2" size="small" className="menylenker-overskrift">
-                                  {columnNode.displayName}
-                              </Heading>
-                              <ul>
-                                  <FooterLenker node={columnNode} />
-                              </ul>
-                          </div>
-                      ))
-                    : [...Array(3)].map((_, index) => (
-                          <div
-                              className={`menylenker-seksjon ${!index ? 'venstre' : index === 2 ? 'hoyre' : 'midt'}`}
-                              key={index}
-                          >
-                              <LinksLoader id={`footer-link-loader-${index}`} />
-                          </div>
-                      ))}
-                <Arbeidsflatevalg />
             </div>
         </div>
     );
